+++ utils.h
@@ -56,7 +56,7 @@ u32 utils_time_diff_ms(u32 start_us, u32 end_us);
 
 u32 utils_time_diff_us(u32 start_us, u32 end_us);
 
-void utils_error_mode(u8 code);
+void utils_error_mode(i32 code);
 
 void utils_internal_led(bool on);
